--- a/consts/errors.go
+++ b/consts/errors.go
@@ -13,6 +13,8 @@ var (
 	ErrInvalidSqueezeLength = errors.New("squeeze length must be a multiple of 243")
 	// ErrInvalidTritsLength gets returned when the trits length are invalid for the given operation.
 	ErrInvalidTritsLength = errors.New("invalid trits length")
+	// ErrInvalidTrytesLength gets returned when the trytes length are invalid for the given operation.
+	ErrInvalidTrytesLength = errors.New("invalid trytes length")
 	// ErrInvalidBytesLength gets returned when the bytes length are invalid for the given operation.
 	ErrInvalidBytesLength = errors.New("invalid bytes length")
 	// ErrInsufficientBalance gets returned when an operation needs a certain amount of balance to fulfill the operation.