--- a/kerl/bigint/bigint.go
+++ b/kerl/bigint/bigint.go
@@ -3,13 +3,11 @@ package bigint
 
 import (
 	"github.com/pkg/errors"
-	"math"
 )
 
 // Errors for bigint package.
 var (
-	ErrUnequallySizedSlices     = errors.New("operation not defined for differently sized slices")
-	ErrSubtractionWithLeftovers = errors.New("could not subtract without leftovers")
+	ErrUnequallySizedSlices = errors.New("operation not defined for differently sized slices")
 )
 
 // MustAdd adds the given big ints together.
@@ -21,9 +19,7 @@ func MustAdd(b []uint32, rh []uint32) {
 	carry := false
 
 	for i := range b {
-		v, c := FullAdd(b[i], rh[i], carry)
-		b[i] = uint32(v)
-		carry = c
+		b[i], carry = FullAdd(b[i], rh[i], carry)
 	}
 }
 
@@ -36,13 +32,7 @@ func MustSub(b []uint32, rh []uint32) {
 	noborrow := true
 
 	for i := range b {
-		v, c := FullAdd(b[i], ^rh[i], noborrow)
-		b[i] = uint32(v)
-		noborrow = c
-	}
-
-	if !noborrow {
-		panic(ErrSubtractionWithLeftovers)
+		b[i], noborrow = FullAdd(b[i], ^rh[i], noborrow)
 	}
 }
 
@@ -51,6 +41,12 @@ func Not(b []uint32) {
 	for i := range b {
 		b[i] = ^b[i]
 	}
+}
+
+// IsNegative checks whether the given big int represents a negative number in two's complement.
+func IsNegative(b []uint32) bool {
+	ms := b[len(b)-1]
+	return (ms >> 31) != 0
 }
 
 // IsNull checks whether the given big int value is null.
@@ -69,55 +65,46 @@ func MustCmp(lh, rh []uint32) int {
 		panic(ErrUnequallySizedSlices)
 	}
 
-	// put LSB first
-	rlh := make([]uint32, len(lh))
-	copy(rlh, lh)
-	ReverseU(rlh)
-
-	rrh := make([]uint32, len(rh))
-	copy(rrh, rh)
-	ReverseU(rrh)
-
-	for i := range rlh {
+	for i := len(lh) - 1; i >= 0; i-- {
 		switch {
-		case rlh[i] < rrh[i]:
+		case lh[i] < rh[i]:
 			return -1
-		case rlh[i] > rrh[i]:
+		case lh[i] > rh[i]:
 			return 1
 		}
 	}
 	return 0
 }
 
-// AddSmall adds a small number to a big int and returns the index of the last carry over.
+// AddSmall adds a uint32 to a big int and returns the highest index that was changed.
 func AddSmall(b []uint32, a uint32) int {
 	v, carry := FullAdd(b[0], a, false)
-	b[0] = uint32(v) // uint is at least 32 bit
-
-	var i int
-	for i = 1; carry; i++ {
-		vi, c := FullAdd(b[i], 0, carry)
-		b[i] = uint32(vi)
-		carry = c
+	b[0] = v
+	if !carry {
+		return 0
 	}
-
-	return i
+	for i := 1; i < len(b); i++ {
+		b[i], carry = FullAdd(b[i], 0, carry)
+		if !carry {
+			return i
+		}
+	}
+	return len(b)
 }
 
-// FullAdd adds left and right together and whether it overflowed.
-func FullAdd(lh, rh uint32, carry bool) (uint, bool) {
-	v, c1 := AddWithOverflow(lh, rh)
+// FullAdd returns the sum and whether the operation overflowed.
+func FullAdd(lh, rh uint32, carry bool) (uint32, bool) {
+	v, c1 := addCarry(lh, rh)
 	var c2 bool
 	if carry {
-		v, c2 = AddWithOverflow(uint32(v), 1)
+		v, c2 = addCarry(v, 1)
 	}
-
 	return v, c1 || c2
 }
 
-// AddWithOverflow returns left hand + right hand and whether it overflowed.
-func AddWithOverflow(lh, rh uint32) (uint, bool) {
-	return uint(lh + rh), lh > math.MaxUint32-rh
+func addCarry(lh, rh uint32) (uint32, bool) {
+	sum := lh + rh
+	return sum, sum < lh
 }
 
 // Reverse reverses the given byte slice.