# Description

This PR improves the binary-ternary conversion used for Kerl, namely `KerlTritsToBytes` and `KerlBytesToTrits`.

- Previously, this was done with a trits<->bytes conversion, leading to uint64 multiplication/divisions for each of the 243 trits. This has been changed to a trytes<->bytes conversion that only needs to be performed for each of the 81 trytes. If necessary, these trytes can then be converted to trits using a LUT without any divisions.
- Introduce `KerlTrytesToBytes` and `KerlBytesToTrytes` when trytes are used directly.
- Adapt `AbsorbTrytes` and `SqueezeTrytes` to work on those functions directly.
- Optimize a few hot functions in `bigint` for speed.
- Change all usages of `Absorb` and `Squeeze` that included a trytes to trits conversion to the corresponding `AbsorbTrytes` and `SqueezeTrytes`.
- Add tests for new tryte-based functions.

On the tested machine this leads to a 2x - 3x speedup for all Kerl related functions. This has been tested with a few artificial benchmarks:
#### Before
```
BenchmarkKerlBytesToTrits-8    	  212587	      5089 ns/op
BenchmarkKerlTritsToBytes-8    	  521274	      2348 ns/op
BenchmarkKerl-8                	  139332	      8144 ns/op
BenchmarkKerlTrytes-8          	  117549	     10121 ns/op
BenchmarkGenerateAddress-8     	      97	  12075147 ns/oo
BenchmarkValidateSignature-8   	     194	   6050952 ns/op
```
#### After
```
BenchmarkKerlBytesToTrits-8    	  757398	      1366 ns/op
BenchmarkKerlTritsToBytes-8    	 1312630	       920 ns/op
BenchmarkKerl-8                	  328953	      3282 ns/o
BenchmarkKerlTrytes-8          	  324426	      3248 ns/op
BenchmarkGenerateAddress-8     	     238	   4918785 ns/op
BenchmarkValidateSignature-8   	     452	   2526877 ns/op
```
## Type of change

- [x] Enhancement (non-breaking change which adds functionality)

# How Has This Been Tested?

- [x] Tests in `iota.go` project test suite
- [x] Additional test cases for tryte based functions
- [x] Benchmarks to verify performance improvement

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have written example code according to the contribution guidelines
- [x] I have added tests using ginkgo that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes